Button navigation and Component page initalization
The buttons on each page direct the user to either the home page to another page on the site.  Initalized the format of the component page and included
functional buttons in the header.  Still need to fix search bar to attempt
to move it to the upper left of the page below the header.