Establish baseline webpage for FAQ Tool
Porting over the existing webpages from the main to current branch.  This branch will contain the webpages that have hardcoded css without any framework used to assist in styling the html pages.  Can be used as a learning tool or as an original framework once edits are committed to the main branch.