Add search bar and moved buttons to navbar. Fixed Accordion.
Switched search bar template for a more functional search bar that requires testing.  Moved buttons to the navbar for a consistent look.  Bolded the words and added spacing between the buttons for more clarity as well.  Added the cosmetic changes to the other pages as well.

Fixed the common questions page so that the accordion is always showing now.