Focus on main page format to include background image
Quick edit to common question page to adjust button highlight to highlight the correct page it is currently on.

Moved search bar to the navbar to keep it consistent across all pages.  Added and formatted text.  Resized the image to fit the entire page and make it more readable. Commented out css lines that for certain were not used.  Added inline css to main html/php page to make up for lost formatting.